feat udistrital/arka_cliente#854: Hallazgos
- Aprobación de entrada movido a archivo aparte.
